server/requirements.txt
flask
python-dotenv
finnhub-python
flask-jwt-extended